consolidate time-ranges and format-time into time
