+++ src/js/control-bar/progress-control/seek-bar.js
@@ -6,7 +6,7 @@ import Component from '../../component.js';
 import {IS_IOS, IS_ANDROID} from '../../utils/browser.js';
 import * as Dom from '../../utils/dom.js';
 import * as Fn from '../../utils/fn.js';
-import {formatTime} from '../../utils/format-time.js';
+import {formatTime} from '../../utils/time.js';
 import {silencePromise} from '../../utils/promise';
 import keycode from 'keycode';
 import document from 'global/document';
