+++ src/js/utils/time.js
@@ -1,6 +1,6 @@
 /**
- * @file time-ranges.js
- * @module time-ranges
+ * @file time.js
+ * @module time
  */
 import window from 'global/window';
 
@@ -28,10 +28,10 @@ import window from 'global/window';
  * @property {number} length
  *           The number of time ranges represented by this object.
  *
- * @property {module:time-ranges~TimeRangeIndex} start
+ * @property {module:time~TimeRangeIndex} start
  *           Returns the time offset at which a specified time range begins.
  *
- * @property {module:time-ranges~TimeRangeIndex} end
+ * @property {module:time~TimeRangeIndex} end
  *           Returns the time offset at which a specified time range ends.
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/API/TimeRanges
@@ -144,3 +144,90 @@ export function createTimeRanges(start, end) {
 }
 
 export { createTimeRanges as createTimeRange };
+
+/**
+ * Format seconds as a time string, H:MM:SS or M:SS. Supplying a guide (in
+ * seconds) will force a number of leading zeros to cover the length of the
+ * guide.
+ *
+ * @private
+ * @param  {number} seconds
+ *         Number of seconds to be turned into a string
+ *
+ * @param  {number} guide
+ *         Number (in seconds) to model the string after
+ *
+ * @return {string}
+ *         Time formatted as H:MM:SS or M:SS
+ */
+const defaultImplementation = function(seconds, guide) {
+  seconds = seconds < 0 ? 0 : seconds;
+  let s = Math.floor(seconds % 60);
+  let m = Math.floor(seconds / 60 % 60);
+  let h = Math.floor(seconds / 3600);
+  const gm = Math.floor(guide / 60 % 60);
+  const gh = Math.floor(guide / 3600);
+
+  // handle invalid times
+  if (isNaN(seconds) || seconds === Infinity) {
+    // '-' is false for all relational operators (e.g. <, >=) so this setting
+    // will add the minimum number of fields specified by the guide
+    h = m = s = '-';
+  }
+
+  // Check if we need to show hours
+  h = (h > 0 || gh > 0) ? h + ':' : '';
+
+  // If hours are showing, we may need to add a leading zero.
+  // Always show at least one digit of minutes.
+  m = (((h || gm >= 10) && m < 10) ? '0' + m : m) + ':';
+
+  // Check if leading zero is need for seconds
+  s = (s < 10) ? '0' + s : s;
+
+  return h + m + s;
+};
+
+// Internal pointer to the current implementation.
+let implementation = defaultImplementation;
+
+/**
+ * Replaces the default formatTime implementation with a custom implementation.
+ *
+ * @param {Function} customImplementation
+ *        A function which will be used in place of the default formatTime
+ *        implementation. Will receive the current time in seconds and the
+ *        guide (in seconds) as arguments.
+ */
+export function setFormatTime(customImplementation) {
+  implementation = customImplementation;
+}
+
+/**
+ * Resets formatTime to the default implementation.
+ */
+export function resetFormatTime() {
+  implementation = defaultImplementation;
+}
+
+/**
+ * Delegates to either the default time formatting function or a custom
+ * function supplied via `setFormatTime`.
+ *
+ * Formats seconds as a time string (H:MM:SS or M:SS). Supplying a
+ * guide (in seconds) will force a number of leading zeros to cover the
+ * length of the guide.
+ *
+ * @example  formatTime(125, 600) === "02:05"
+ * @param    {number} seconds
+ *           Number of seconds to be turned into a string
+ *
+ * @param    {number} guide
+ *           Number (in seconds) to model the string after
+ *
+ * @return   {string}
+ *           Time formatted as H:MM:SS or M:SS
+ */
+export function formatTime(seconds, guide = seconds) {
+  return implementation(seconds, guide);
+}
