+++ test/unit/seek-to-live.test.js
@@ -2,7 +2,7 @@
 import TestHelpers from './test-helpers.js';
 import sinon from 'sinon';
 import { computedStyle } from '../../src/js/utils/dom.js';
-import { createTimeRange } from '../../src/js/utils/time-ranges.js';
+import { createTimeRange } from '../../src/js/utils/time.js';
 
 QUnit.module('SeekToLive', {
   beforeEach() {
